Add autoSaveAction lazy closure var customization point
That is, when disappearing from the view, the settor
of that closure could infer from the passed
PDFAnnotationController whether there are things
to save, and if so, decide where/what to do or
just ignore etc.

Default behaviour is as before, just assigns the
annotations out of the annotationController to the
PDFDocument itself, then call save() on that
document to archiveRootObject.